shared memory and db handler split, serial reader changed a bit
